[Feat/#4] StopwatchView 컴포넌트화
