+++ PlanFit/PlanFit/Presentation/Workout/View/WorkoutView.swift
@@ -13,15 +13,7 @@ class WorkoutView: UIView {
     
     // MARK: - UIComponent
     
-    let currentTimeView = UIView()
-    
     let collectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewLayout())
-    
-    private let currentTimeOnAirDot = UIImageView()
-    
-    private let currentTimeLabel = UILabel()
-    
-    private let currentTimePlayImage = UIImageView()
     
     private lazy var breakTimerButton = UIButton()
     
@@ -47,25 +39,6 @@ class WorkoutView: UIView {
 private extension WorkoutView {
     func setUI() {
         backgroundColor = .gray08BG
-        
-        currentTimeView.do {
-            $0.backgroundColor = .none
-            $0.layer.borderColor = UIColor.mainGreen.cgColor
-            $0.layer.borderWidth = 1
-            $0.layer.cornerRadius = 15
-        }
-        
-        currentTimeOnAirDot.do {
-            $0.image = UIImage(resource: .onairDot)
-        }
-        
-        currentTimeLabel.do {
-            $0.setText("00:00:37", font: .subtitle01, color: .gray01)
-        }
-        
-        currentTimePlayImage.do {
-            $0.image = UIImage(resource: .play)
-        }
         
         collectionView.do {
             $0.showsVerticalScrollIndicator = false
@@ -98,35 +71,10 @@ private extension WorkoutView {
     }
     
     func setViewHierarchy() {
-        currentTimeView.addSubviews(currentTimeOnAirDot, currentTimeLabel, currentTimePlayImage)
-        
         addSubviews(collectionView, breakTimerButton, completeSetButton)
     }
     
     func setAutoLayout() {
-        currentTimeView.snp.makeConstraints {
-            $0.width.equalTo(133)
-            $0.height.equalTo(30)
-        }
-        
-        currentTimeOnAirDot.snp.makeConstraints {
-            $0.leading.equalToSuperview().offset(10)
-            $0.centerY.equalToSuperview()
-            $0.width.height.equalTo(16)
-        }
-        
-        currentTimeLabel.snp.makeConstraints {
-            $0.leading.equalTo(currentTimeOnAirDot.snp.trailing).offset(1)
-            $0.centerY.equalToSuperview()
-        }
-        
-        currentTimePlayImage.snp.makeConstraints {
-            $0.leading.equalTo(currentTimeLabel.snp.trailing).offset(1)
-            $0.trailing.equalToSuperview().offset(-11)
-            $0.centerY.equalToSuperview()
-            $0.width.height.equalTo(16)
-        }
-        
         collectionView.snp.makeConstraints {
             $0.top.leading.trailing.equalTo(safeAreaLayoutGuide)
             $0.bottom.equalTo(breakTimerButton.snp.top)
